pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = "Backend part for unitrip.ru"
authors = ["Exilon"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
uvicorn = "^0.20.0"
fastapi = "^0.87.0"
alembic = "^1.8.1"
asyncpg = "^0.27.0"
python-dotenv = "^0.21.0"
sqlalchemy = "^1.4.44"
pydantic = {extras = ["email"], version = "^1.10.2"}
psycopg2 = "^2.9.5"
fastapi-jwt-auth = {extras = ["asymmetric"], version = "^0.5.0"}
loguru = "^0.6.0"
passlib = {extras = ["context"], version = "^1.7.4"}
python-jose = "^3.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
